Made WallGoResults.__init__() just a pass. Moved dataclasses to WallGoTypes.
